src/scripts/requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
bs4==0.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.7.0
django-environ==0.13.0
djangorestframework==3.17.1
gevent==26.4.0
greenlet==3.4.0
gunicorn==25.3.0
idna==3.11
lxml==6.0.2
numpy==2.4.4
packaging==26.2
pandas==3.0.2
playwright==1.58.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.12
pycparser==3.0
pyee==13.0.1
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
websocket-client==1.9.0
whitenoise==6.12.0
xlrd==2.0.2
zope.event==6.2
zope.interface==8.4
